Create modularized common config file
